Move KCP CLI out from KEB
