+++ components/kyma-environment-broker/Gopkg.toml
@@ -79,23 +79,3 @@
   go-tests = true
   unused-packages = true
   non-go = true
-
-[[constraint]]
-  name = "github.com/spf13/cobra"
-  version = "1.0.0"
-
-[[constraint]]
-  name = "github.com/spf13/viper"
-  version = "1.7.1"
-
-[[constraint]]
-  name = "github.com/int128/kubelogin"
-  version = "1.21.0"
-
-[[constraint]]
-  name = "github.com/spf13/pflag"
-  version = "1.0.5"
-
-[[constraint]]
-  name = "golang.org/x/oauth2"
-  revision = "5d25da1a8d43b66f2898c444f899c7bcfd6a407e"
