--- components/kyma-environment-broker/Makefile.cli
@@ -1,41 +1,0 @@
-#
-# Make targets to build kcp CLI artifacts
-# source code verification (make verify) related targets (go-test check-imports check-fmt errcheck) are covered by the main Makefile
-#
-
-ifndef CLI_VERSION
-	CLI_VERSION = ${shell git describe --tags --always}
-endif
-ifndef ARTIFACTS
-	ARTIFACTS = .
-endif
-
-CLI_FLAGS = -ldflags '-s -w -X github.com/kyma-project/control-plane/components/kyma-environment-broker/cmd/cli/command.Version=$(CLI_VERSION)'
-
-.PHONY: release
-release: resolve build
-
-.PHONY: resolve
-resolve:
-	dep ensure -vendor-only -v
-	dep status -v
-
-
-.PHONY: build
-build: build-windows build-linux build-darwin
-
-.PHONY: build-windows
-build-windows:
-	CGO_ENABLED=0 GOOS=windows GOARCH=amd64 go build -o $(ARTIFACTS)/kcp.exe $(CLI_FLAGS) ./cmd/cli
-
-.PHONY: build-linux
-build-linux:
-	CGO_ENABLED=0 GOOS=linux GOARCH=amd64 go build -o $(ARTIFACTS)/kcp-linux $(CLI_FLAGS) ./cmd/cli
-
-.PHONY: build-darwin
-build-darwin:
-	CGO_ENABLED=0 GOOS=darwin GOARCH=amd64 go build -o $(ARTIFACTS)/kcp-darwin $(CLI_FLAGS) ./cmd/cli
-
-.PHONY: docs
-docs:
-	go run ./cmd/cli/gendocs/gendocs.go
