+++ tools/cli/cmd/main/main.go
@@ -6,8 +6,8 @@ import (
 	"os/signal"
 	"syscall"
 
-	"github.com/kyma-project/control-plane/components/kyma-environment-broker/cmd/cli/command"
-	"github.com/kyma-project/control-plane/components/kyma-environment-broker/cmd/cli/logger"
+	"github.com/kyma-project/control-plane/tools/cli/pkg/command"
+	"github.com/kyma-project/control-plane/tools/cli/pkg/logger"
 )
 
 func main() {
