+++ tools/cli/pkg/command/kubeconfig.go
@@ -9,9 +9,9 @@ import (
 	"github.com/pkg/errors"
 
 	"github.com/kyma-project/control-plane/components/kubeconfig-service/pkg/client"
-	"github.com/kyma-project/control-plane/components/kyma-environment-broker/cmd/cli/credential"
-	"github.com/kyma-project/control-plane/components/kyma-environment-broker/cmd/cli/logger"
 	"github.com/kyma-project/control-plane/components/kyma-environment-broker/common/runtime"
+	"github.com/kyma-project/control-plane/tools/cli/pkg/credential"
+	"github.com/kyma-project/control-plane/tools/cli/pkg/logger"
 	"github.com/spf13/cobra"
 	"gopkg.in/yaml.v2"
 )
