+++ tools/cli/pkg/command/login.go
@@ -3,7 +3,7 @@ package command
 import (
 	"errors"
 
-	"github.com/kyma-project/control-plane/components/kyma-environment-broker/cmd/cli/logger"
+	"github.com/kyma-project/control-plane/tools/cli/pkg/logger"
 	"github.com/spf13/cobra"
 )
