+++ tools/cli/pkg/command/orchestration.go
@@ -9,9 +9,9 @@ import (
 
 	"github.com/pkg/errors"
 
-	"github.com/kyma-project/control-plane/components/kyma-environment-broker/cmd/cli/logger"
-	"github.com/kyma-project/control-plane/components/kyma-environment-broker/cmd/cli/printer"
 	"github.com/kyma-project/control-plane/components/kyma-environment-broker/common/orchestration"
+	"github.com/kyma-project/control-plane/tools/cli/pkg/logger"
+	"github.com/kyma-project/control-plane/tools/cli/pkg/printer"
 	"github.com/spf13/cobra"
 )
