+++ tools/cli/pkg/command/runtime.go
@@ -1,9 +1,9 @@
 package command
 
 import (
-	"github.com/kyma-project/control-plane/components/kyma-environment-broker/cmd/cli/logger"
-	"github.com/kyma-project/control-plane/components/kyma-environment-broker/cmd/cli/printer"
 	"github.com/kyma-project/control-plane/components/kyma-environment-broker/common/runtime"
+	"github.com/kyma-project/control-plane/tools/cli/pkg/logger"
+	"github.com/kyma-project/control-plane/tools/cli/pkg/printer"
 	"github.com/pkg/errors"
 	"github.com/spf13/cobra"
 )
